DRAFT - Mutual Information virtual space truncation (#155)
# Summary

Localizer have been reorganised so that there are seperate base classes for virtual and occupied localizers. Concentric localization has been moved out of the occupied SPADE localizer, enabling its use with any occupied localization method.

The structure is now as follows.

```
nbed
-localizers
|- virtual
  |-concentric
|- occupied
 |- spade
 |-pyscf
```
Additionally functionality has been added to use the MI to truncate the active space. A significant complication to this is that there should be an energetic correction to accunt for this, but after a fair bit of searching no obvious answer has come up. A possible minimum solution could be to take the difference in the HF energy for systems with different numbers of orbitals.

Foolishly, this PR also includes updates to the repo dependency management, moving from poetry to uv. There is still some work to be done, with documenting this and running CICD. This will follow in another (hopefully more suitable scoped PR)