add test for emb_fci (#160)
The `NbedDriver._run_emb_FCI` function was failing for open shelled systems. The below explanation comes from issue #157.


So theres a couple of issues that led to this, stemming from:
- The logic of PBE
- Numpy magic
- Assumed behaviour in PySCF.FCI

# Nbed
To account for the effect of the environment on the active region, we add a constant effective potential to the one-electron hamiltonian. For a PySCF.scf object, we do this by updating the `get_hcore()` function. You can see this in the `mu_embed` and `huzinaga_embed` functions of `NbedDriver`, we basically substitude a lambda for the original function and add the constant term.

```python
original_hcore = embedded_scf.get_hcore
embedded_scf.get_hcore = lambda *args: original_hcore(*args) + v_embedding
```

# Numpy Magic
The reason this ends up causing an error is because we can have a different number of dimensions for hcore and v_emb. The one-electron hamiltonian doesn't include any effects from spin, so for a system with M spatial orbitals the shape of the output expected from `get_hcore()` is `(M,M)`. The embedding potential can be different for each spin (we currently always have an active region with an even number of electrons, so what happens when the whole molecule is open-shell?) v_emb ends up having shape (2,M,M).

It isn't clear this is happening in the code above, numpy very helpfully broadcasts h_core to the correct shape, but it does so silently. This kind of helpfullness **could** propagate down every layer of the logic, except that by default, numpy defines the transpose `ndarray([...]).T` of an N-dimensional matrix as a swap of the first two axes. In fact it is equivalent to `np.swapaxes(<array>, 0,1)`. So when the transpose is applied, we end up with an array of shape (M,2,M), which numpy doesn't know how to broadcast. That's the bit that actaully raises the error we see.

# PySCF
When it comes time to use `pyscf.ci.FCI`, the very first thing that object does is [check what type of input it got](https://github.com/pyscf/pyscf/blob/142756584ed3963728cc1f6823286f4605811a6c/pyscf/fci/__init__.py#L99). In the example above, we're passing it an `scf.UHF`. This should be the type we need, because it allows treating the two spins differently from eachother. For example, if you run `get_veff()` on the scf object we are passing to FCI classs, you'll find its (2,M,M). This is why we have that shape for `v_emb`, to keep the interface consistent with PySCF. 

However, PySCF assumes that [even for a UHF, the h_core is 2D](https://github.com/pyscf/pyscf/blob/142756584ed3963728cc1f6823286f4605811a6c/pyscf/fci/__init__.py#L159). It calculate a one-electron hamiltonian `h1e` and passes that as a default the `FCI.kernel()` function. For a UHF it separates out the alpha and beta molecular orbitals, and applies a dot product with `hcore`, to get `[h1e_a, h1e_b]`. At this point __numpy broadcasts the term__ for each spin to (2,M,M)! So now we know why we have `h1e_a` with shape (11,2,11). 

# The Solution

Since the value is just passed to the `kernel` function, we can define it ourselves
```python
import numpy as np
from functools import reduce
mo = mol_driver.embedded_scf.mo_coeff
hcore = mol_driver.embedded_scf.get_hcore()
h1e = [reduce(np.dot, (mo[0].T, hcore[0], mo[0])), reduce(np.dot, (mo[1].T, hcore[1], mo[1]))]
nbed_embed_fci.kernel(h1e=h1e)
```

I'll update the `_run_emb_FCI` function to do that, and add some tests. For the meantime you can add that to your own code. Hopefully that explanation was easy enough to follow. I wanted to do a little more than just provide a code snippet since it wouldn't be clear why that would be