Open shell embedded region (#168)
Sets the default SCF methods for both the whole system and the embedded region to be unrestricted.

Localization is then performed on each spin individually. As a result, different numbers of orbitals may be localized between spins.

Additionally, the spin and charge of the whole system will be passed on to the embedded region.

Much of the code for restricted methods remains in the codebase and unchanged, so in future if the need emerges it should be possible to extend the current set up to:
- use one localization for both spins
- restrict the embedded region only
- restrict the environment only