Embedding non-zero spin and charge (#170)
A previous PR moved the codebase over to unrestricted SCF methods by default.

This PR makes changes to the localizers and local scf functions which:

- enforces localization of MOs to be consistent between spins

When different occupancies exist between spins, this results in occupied c matrices which do not match in size or in incuded MOs.

To generate matching occupied c matrices, the localize_spin functions are called again, but using the sum of occupations between spins. The c matrices from the second *consistent* runs are combined with the MO indices from the individual runs.

- carries over localization of electrons to create a local molecule with spin and charge determined by localization procedures and not config.

Using the MO indices from the individual spin localizations gives the number of electrons which should be in each spin of the embedded system.
The spin of the embedded system is calculated as the difference n_alpha - n_beta.

When deleting the environment, we combine the enviro MO indices from each of the individual spin localizations. The union of these sets gives the number of environment MOs to remove.