Simpler localized system object (#184)
In preparation for handling Restricted open, this PR includes some simplifications to the LocalizedSystem dataclass.

The idea is that in future we will want to handle spin localizations of different numbers of electrons in each spin, without having to explicitly check for this downstream.

This should make it easier to handle the change in array dimension between MO coefficient matrix and occupancies for, ROKS, RKS, UKS respectively.
property\method 	RKS 	ROKS 	UKS
Occupancy 	(n) 	(2,n) 	(2,n)
MO Coeff 	(M,M) 	(M,M) 	(2,M,M)

This gets extra complicated when we allow any combination of R,RO,U as each of input and output, with variable number of electrons.

To reduce the variability in the objects that are being stored and tracked,

    occupancy arrays are now boolean arrays rather than arrays of indices. This preserves the shape (n) for each spin.
    Rather than pass c_active and c_enviro, we instead pass density matrices, which preserve the shape (M,M) for each dm, rather than ( M , n a c t ) .