Typing with pyrefly (#186)
Adds typing with pyrefly, ignoring tests/. The hope is that this will make it easier to make the changes needed to support restricted open.

There are quite a few ignores throughout the code, as pyscf has very little if any typing. The most common cause for ignoring typing is that SCF methods do not have an attribute for mo_coeff, mo_occ etc, seemingly not until after kernel has been run.

type checking with ty brings up a whole different set of issues, but this will do for now.